[REFACTOR] : created DictEntity
