+++ MLS/MLS/Models/Dictionary/DictMap.swift
@@ -1,5 +1,5 @@
 //
-//  DictionaryMap.swift
+//  DictMap.swift
 //  MLS
 //
 //  Created by JINHUN CHOI on 2024/02/07.
@@ -8,15 +8,15 @@
 import Foundation
 
 // MARK: - Dev 1.0.0
-struct DictionaryMap: Codable {
-    var code: String
-    var name: String
-    var monsters: [DictionaryNameDescription?] // 이름 , 갯수
-    var npcs: [String?]
-}
+//struct DictionaryMap: Codable {
+//    var code: String
+//    var name: String
+//    var monsters: [DictionaryNameDescription?] // 이름 , 갯수
+//    var npcs: [String?]
+//}
 
 // MARK: - Dev 1.0.1
-struct DictMap: Sqlable {
+struct DictMap: DictEntity, Sqlable {
     var code: String
     var name: String
     var monsters: [DictionaryNameDescription] // 이름 , 갯수
