+++ MLS/MLS/Models/Dictionary/DictNPC.swift
@@ -1,5 +1,5 @@
 //
-//  DictionaryNPC.swift
+//  DictNPC.swift
 //  MLS
 //
 //  Created by JINHUN CHOI on 2024/02/07.
@@ -8,14 +8,14 @@
 import Foundation
 
 // MARK: - Dev 1.0.0
-struct DictionaryNPC: Codable {
-    var code: String
-    var name: String
-    var quests: [DictionaryNameDescription?]
-}
+//struct DictionaryNPC: Codable {
+//    var code: String
+//    var name: String
+//    var quests: [DictionaryNameDescription?]
+//}
 
 // MARK: - Dev 1.0.1
-struct DictNPC: Sqlable {
+struct DictNPC: DictEntity, Sqlable {
     var code: String
     var name: String
     var maps: [String]
